sl2.8-1.3.sql: add a cast ::int to a null, for psql to be able to match the function signature.
git-svn-id: svn://svn.code.sf.net/p/ledger-smb/code/branches/1.3@6143 4979c152-3d1c-0410-bac9-87ea11338e46

+++ sql/upgrade/sl2.8-1.3.sql
@@ -26,10 +26,11 @@ INSERT INTO account_heading(id, accno, description)
 SELECT id, accno, description
   FROM sl28.chart WHERE charttype = 'H';
 
-SELECT account_save(id, accno, description, category, gifi_accno, NULL, contra, 
-                    CASE WHEN link like '%tax%' THEN true ELSE false END, 
+SELECT account_save(id, accno, description, category, gifi_accno, NULL::int,
+                    contra,
+                    CASE WHEN link like '%tax%' THEN true ELSE false END,
                     string_to_array(link,':'))
-  FROM sl28.chart 
+  FROM sl28.chart
  WHERE charttype = 'A';
 
 delete from account_link where description = 'CT_tax';
